p2p: Separate address lookup service as nameserver
